requirements.txt
matplotlib==2.0.2
numpy==1.14.0
pyserial==3.4
wxPython==1.11.0
